Refactor the statistics view and add the logic to use the custom index pattern
