+++ public/controllers/management/components/management/statistics/hooks/index.ts
@@ -1,0 +1,14 @@
+/*
+ * Wazuh app - React component for building Remoted dashboard
+ *
+ * Copyright (C) 2015-2020 Wazuh, Inc.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * Find more information about this on the LICENSE file.
+ */
+
+export { useBuildStatisticsVisualizations } from './use-build-statistics-visualizations';
